use the tag name instead of the id in the tag show path
we could do this by using the tag.name method when building the link.
But to prep for Rails, I used the to_param method from ActiveRecord. The
nice thing about this is that it's a responsibility of the tag to tell
the view how to represent it in a url (ie. not a view concern). So using
to_param is not only Railsy, it's also better OO.