add a login form
I've chosen to use a non-RESTful route for login. There are really 2
schools of thought on login/logout routing.

1. non-RESTful, so the typical routes: `get "/login"` for the login form,
`post "/login"` validates credentials from the form and stores the
user_id in the session, and `get "/logout"` to remove the user_id from
the session.
2. RESTful. In this case, the "resource" being created and destroyed is
a session. While the session isn't a resource we store in the database,
it is something we can think of as a resource in some sense. The RESTful
routes are typically: `get "/session/new"` for the login form, `post
"/session"` validates the credentials from the form and stores the
user_id in the session, and `delete "/session"` to remove the user_id
from the session.

You can see that in either case we're doing the same work, and in fact
the code will look exactly the same. The only difference is the verb and
path combinations.

One other difference you'll notice in the RESTFUL routes is the use of
"session" rather than "sessions". This is considered a singular
resource. Think about it, if we didn't reference this as a singular
resource, we'd typically need to provide an "id" in the "destroy" action
to reference which session we were destroying.
+