allow only the recipe's chef to edit and update it
notice that not only did I change the view, so that we no longer show
the edit link to other users, I also ensure that both the edit and
update actions in my controller load the given recipe from the
current_user's list of recipes. If the current_user isn't the owner of
the recipe, `find` will raise an RecordNotFoundError, which is perfect,
because the client was probably fiddling with the URL to try to access
another user's recipes anyway. In development mode, you'd see this error
in the browser. In production, this error will display as a 404 (which
is the HTTP status for Not Found!).